MSG-30 Server bzl refactor worker extracted
