+++ src/main/java/Server/ServerMain.java
@@ -1,43 +1,23 @@
 package Server;
 
 import java.io.IOException;
-import java.io.OutputStream;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.Date;
 
 public class ServerMain {
     public static void main(String[] args) {
         int port = 1400;
         try {
             ServerSocket serverSocket = new ServerSocket(port);
-            while(true) {
+            while (true) {
                 System.out.println("Accepting client connection");
                 Socket clientSocket = serverSocket.accept();
                 System.out.println("Connection accepted from " + clientSocket);
-                Thread t = new Thread() {
-                    @Override
-                    public void run() {
-                        try {
-                            handleClientSocket(clientSocket);
-                        } catch (IOException | InterruptedException e) {
-                            e.printStackTrace();
-                        }
-                    }
-                };
-                t.start();
+                ServerWorker worker = new ServerWorker(clientSocket);
+                worker.start();
             }
         } catch (IOException e) {
             e.printStackTrace();
         }
-    }
-
-    private static void handleClientSocket(Socket clientSocket) throws IOException, InterruptedException {
-        OutputStream outputStream = clientSocket.getOutputStream();
-        for (int i = 0 ; i < 10 ; ++i) {
-            outputStream.write(("Current time: " + new Date() + "\n").getBytes());
-            Thread.sleep(1000);
-        }
-        clientSocket.close();
     }
 }
